pyproject.toml
[tool.poetry]
name = "is-replicant"
version = "0.1.0"
description = "AI-generated text detection API"
authors  = ["Raniro Coelho <ranirocoelho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
pydantic = "^2.8.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
loguru = "^0.7.2"
nltk = "^3.8.1"
scikit-learn = "^1.5.1"
pandas = "^2.2.2"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
alembic = "^1.13.2"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
